Edit Coordinate - Deg-Min-Sec
git-svn-id: svn://svn.code.sf.net/p/droidcachebox/code/trunk@54 5890f383-b09f-485f-8f59-22bc967d67fd

+++ src/de/droidcachebox/Geocaching/Coordinate.java
@@ -145,10 +145,10 @@ public class Coordinate implements Serializable {
         {
           // Decimal - Minute - Second
           lat = Integer.valueOf(llat.get(0));
-          lat += Integer.valueOf(llat.get(1)) / 60;
+          lat += Double.valueOf(llat.get(1)) / 60;
           lat += Double.valueOf(llat.get(2)) / 3600;
           lon = Integer.valueOf(llon.get(0));
-          lon += Integer.valueOf(llon.get(1)) / 60;
+          lon += Double.valueOf(llon.get(1)) / 60;
           lon += Double.valueOf(llon.get(2)) / 3600;
         }
       }
